+++ deltafi-common/src/main/java/org/deltafi/common/content/ContentStorageService.java
@@ -21,7 +21,6 @@ import lombok.RequiredArgsConstructor;
 import org.deltafi.common.storage.s3.ObjectReference;
 import org.deltafi.common.storage.s3.ObjectStorageException;
 import org.deltafi.common.storage.s3.ObjectStorageService;
-import org.springframework.stereotype.Service;
 
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
@@ -29,7 +28,6 @@ import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
-@Service
 @RequiredArgsConstructor
 public class ContentStorageService {
     public static final String CONTENT_BUCKET = "storage";