+++ deltafi-common/src/main/java/org/deltafi/common/storage/s3/minio/MinioAutoConfiguration.java
@@ -1,0 +1,42 @@
+/**
+ *    DeltaFi - Data transformation and enrichment platform
+ *
+ *    Copyright 2022 DeltaFi Contributors <deltafi@deltafi.org>
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+package org.deltafi.common.storage.s3.minio;
+
+import io.minio.MinioClient;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+import org.springframework.context.annotation.Bean;
+
+@AutoConfiguration
+@ConfigurationPropertiesScan
+public class MinioAutoConfiguration {
+    @Bean
+    public MinioClient minioClient(MinioProperties minioProperties) {
+        return MinioClient.builder()
+                .endpoint(minioProperties.getUrl())
+                .credentials(minioProperties.getAccessKey(), minioProperties.getSecretKey()).build();
+    }
+
+    @Bean
+    @ConditionalOnMissingBean
+    public MinioObjectStorageService minioObjectStorageService(MinioClient minioClient,
+                                                               MinioProperties minioProperties) {
+        return new MinioObjectStorageService(minioClient, minioProperties);
+    }
+}